move ctrl and ControlledOperation to op_math submodule
