+++ pennylane/__init__.py
@@ -46,7 +46,7 @@ from pennylane.tracker import Tracker
 from pennylane.io import *
 from pennylane.measurements import density_matrix, measure, expval, probs, sample, state, var
 from pennylane.ops import *
-from pennylane.ops import adjoint
+from pennylane.ops import adjoint, ctrl
 from pennylane.templates import broadcast, layer
 from pennylane.templates.embeddings import *
 from pennylane.templates.layers import *
@@ -63,9 +63,7 @@ from pennylane.transforms import (
     batch_partial,
     cut_circuit,
     cut_circuit_mc,
-    ControlledOperation,
     compile,
-    ctrl,
     cond,
     defer_measurements,
     measurement_grouping,
